requirements.txt
pytest==8.3.4
pylint==3.3.4
black==25.1.0
requests==2.32.3
pandas==2.2.3
sqlalchemy==2.0.38
mysql-connector-python==9.2.0
python-dotenv==1.0.1
click==8.1.8
pytest-mock==3.14.0